+++ p2p/discv5/endpoint_tracker.py
@@ -14,7 +14,7 @@ from eth_utils.toolz import (
     merge,
 )
 
-from p2p.trio_service import (
+from p2p.service import (
     Service,
 )
