+++ p2p/service/__init__.py
@@ -1,0 +1,7 @@
+from .abc import (  # noqa: F401
+    ServiceAPI,
+    ManagerAPI,
+)
+from .base import external_api, as_service, Service  # noqa: F401
+from .exceptions import LifecycleError, DaemonTaskExit, ServiceCancelled  # noqa: F401
+from .trio import TrioManager, background_trio_service  # noqa: F401
